All test passed

+++ vector/Vector.h
@@ -2,7 +2,7 @@
 
 #include <iostream>
 
-using Value = int;
+using Value = double;
 
 //! \brief Класс, реализующий С++ контейнер vector для типа int
 class Vector {
